user crud

+++ tieba-infrastructure/src/main/java/com/tingyu/tieba/mappers/BaMapper.java
@@ -1,6 +1,5 @@
 package com.tingyu.tieba.mappers;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 import com.tingyu.tieba.ba.dataobject.BaDO;
@@ -12,9 +11,9 @@ public interface BaMapper {
 
   public Long create(BaDO baDO);
 
-  public BigDecimal update(BaDO baDO);
+  public Long update(BaDO baDO);
 
-  public BigDecimal delete(@Param("id") BigInteger id);
+  public Long delete(@Param("id") BigInteger id);
 
   public List<BaDO> list(BaDO baDO);
